requirements.txt
# ElevenLabs TTS
elevenlabs==1.8.0

# Audio Processing
pydub==0.25.1
noisereduce==3.0.0
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3

# Configuration & Environment
python-dotenv==1.0.0
pyyaml==6.0.1

# HTTP & Utilities
requests==2.31.0

# Optional but helpful
tqdm==4.66.1